requirements.txt
allure-behave>=2.13.0
behave>=1.2.6
allure-pytest>=2.13.0
allure-python-commons>=2.13.0
Appium-Python-Client>=2.8.1
async-generator>=1.10
atomicwrites>=1.4.0
attrs>=21.2.0
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.11
colorama>=0.4.4
configparser>=5.2.0
crayons>=0.4.0
cryptography>=36.0.0
decorator>=5.1.1
distlib>=0.3.6
docutils>=0.18.1
enum34>=1.1.10
et-xmlfile>=1.1.0
exceptiongroup>=1.1.0
execnet>=1.9.0
filelock>=3.9.0
h11>=0.12.0
idna>=3.3
importlib-metadata>=4.11.1
iniconfig>=1.1.1
lxml>=4.9.1
MouseInfo>=0.1.3
namedlist>=1.8
numpy>=1.21.5
opencv-python>=4.5.4.60
openpyxl>=3.0.9
outcome>=1.1.0
packaging>=21.3
Pillow>=8.4.0
platformdirs>=3.1.0
pluggy>=1.0.0
py>=1.11.0
PyAutoGUI>=0.9.53
PyAutoIt>=0.6.5
pycparser>=2.21
PyGetWindow>=0.0.9
Pygments>=2.10.0
PyMsgBox>=1.0.9
pyOpenSSL>=21.0.0
pyparsing>=3.0.7
pyperclip>=1.8.2
PyRect>=0.2.0
PyScreeze>=0.1.28
PySocks>=1.7.1
pytesseract>=0.3.8
pytest>=7.2.2
pytest-xdist>=3.2.0
python-dateutil>=2.8.2
pytweening>=1.0.4
PyUserInput>=0.1.10
selenium>=4.8.2
six>=1.16.0
sniffio>=1.2.0
sortedcontainers>=2.4.0
tomli>=2.0.1
trio>=0.19.0
trio-websocket>=0.9.2
junit-xml>=1.9